+++ lib/common_widgets/coordinates/gcw_coords/coord_format_inputs/reversewherigo_waldmeister/widget/gcw_coords_reversewherigo_waldmeister.dart
